TilApp: switches from existing class components logic to functional component
